+++ csv_detective/detect_fields/geo/json_geojson/__init__.py
@@ -1,6 +1,7 @@
 import json
 
 PROPORTION = 0.9
+PARENT = "json"
 
 
 def _is(val):