+++ csv_detective/detect_fields/temp/datetime_rfc822/__init__.py
@@ -1,6 +1,7 @@
 import re
 
 PROPORTION = 1
+PARENT = "datetime_aware"
 
 
 def _is(val):