+++ csv_detective/detect_fields/temp/year/__init__.py
@@ -1,4 +1,5 @@
 PROPORTION = 1
+PARENT = "int"
 
 
 def _is(val):